Avoid collision in @Bean name
url-locale is often used in projects that make use of Spring's Web MVC.
WebMvcAutoConfiguration includes a @Bean with name "localeResolver" (with type
org.springframework.web.servlet.LocaleResolver). As we cannot have 2 beans with
the same name any app with both url-locale and Web MVC will fail to start.

By renaming the method the bean is defined with we rename the bean and avoid
the collision.

https://github.com/spring-projects/spring-boot/blob/1395e1eb3251ef8351fdb5eed8f92d4731cd076f/spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/web/servlet/WebMvcAutoConfiguration.java#L301