Add lang request param feature
Consumers of this library  can enable the feature in their configuration using
`url-locale.lang-parameter-enabled=true`.

When enabled, if the request includes a lang parameter the url locale resolver
will attempt to use that language instead of the default language for the url
locale.

The language parameter must be an ISO 639-1 alpha 2 language code. It also must
be supported. The set of supported languages is the set of languges used by the
existing url locales. If the lang parameter does not match one of these
languages the default language for the url locale will be used.